+++ oci/oci_linux.go
@@ -3,6 +3,7 @@
 package oci
 
 import (
+	"fmt"
 	"os"
 	"path/filepath"
 	"syscall"
@@ -16,7 +17,11 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-func (r *Runtime) createContainerPlatform(c *Container, cgroupParent string, pid int) error {
+func createUnitName(prefix string, name string) string {
+	return fmt.Sprintf("%s-%s.scope", prefix, name)
+}
+
+func (r *RuntimeV1) createContainerPlatform(c *Container, cgroupParent string, pid int) error {
 	// Move conmon to specified cgroup
 	if r.cgroupManager == SystemdCgroupsManager {
 		logrus.Debugf("Running conmon under slice %s and unitName %s", cgroupParent, createUnitName("crio-conmon", c.id))
